% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_from_raster.R
\name{add_from_raster}
\alias{add_from_raster}
\title{Adds values of a RasterLayer object to a SpatialPointsDataFrame object}
\usage{
add_from_raster(sptsdf, rstr, varname = "new_data")
}
\arguments{
\item{sptsdf}{A \code{SpatialPointsDataFrame} object.}

\item{rstr}{A \code{RasterLayer} object.}

\item{varname}{The name of the new variable. Defaults to "new_data".}
}
\value{
A \code{SpatialPointsDataFrame} object.
}
\description{
Based on the \code{\link[raster]{extract}} function.
}
\examples{
library(sp)
library(magrittr) # for " \%>\% "

vn <- sptools::gadm("vietnam", "sp", 0)

# With a SpatialPointsDataFrame
value <- vn \%>\%
  spsample(100, "random") \%>\%
  SpatialPointsDataFrame(data.frame(variable = 1:100)) \%>\%
  add_from_raster(srtmVN::getsrtm(), "elevation")

value \%>\%
  slot("data") \%>\%
  head()

# With a SpatialPoints
value <- vn \%>\%
  spsample(100, "random") \%>\%
  SpatialPoints(proj4string = vn@proj4string, bbox = vn@bbox) \%>\%
  add_from_raster(srtmVN::getsrtm(), "elevation")

value \%>\%
  slot("data") \%>\%
  head()
}
