% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variable.R
\name{add_variable_sp}
\alias{add_variable_sp}
\title{Adds variable(s) to a SpatialPolygons}
\usage{
add_variable_sp(sp, df)
}
\arguments{
\item{sp}{A SpatialPolygons object.}

\item{df}{A vector or a data frame.}
}
\value{
A SpatialPolygonsDataFrame
}
\description{
Adds variable(s) to a SpatialPolygons
}
\examples{
vn <- gadm("vietnam", "sp", 1)
vn_nodata <- remove_data_spatialpolygons(vn)
vn_nodata

a <- runif(63)
vn_data <- add_variable_sp(vn_nodata, a)
vn_data

# An example with a data frame:
vn_data2 <- add_variable_sp(vn_nodata,
   data.frame(b = rnorm(63), c = rexp(63)))
vn_data2

}
