% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variable.R
\name{add_variable_spdf}
\alias{add_variable_spdf}
\title{Adds variable(s) to a Spatial*DataFrame}
\usage{
add_variable_spdf(spdf, df)
}
\arguments{
\item{spdf}{A Spatial*DataFrame object}

\item{df}{A vector or a data frame}
}
\value{
An object of the same class than \code{sp}.
}
\description{
Adds variable(s) to a Spatial*DataFrame
}
\examples{
library(magrittr)

vn <- gadm("vietnam", "sp", 1)
vn
vn \%<>\% add_variable_spdf(rnorm(63))
vn

# The name of the new variable will be "df". To prevent that we could choose
# the data frame format:
vn \%<>\% add_variable_spdf(data.frame(normal = rnorm(63)))
vn

# The data frame format allows the additional possibility to add several
# additional variables all at once:
vn \%<>\% add_variable_spdf(
  data.frame(uniform = runif(63), exponential = rexp(63)))
vn
}
