% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variable.R
\name{add_variable_spts}
\alias{add_variable_spts}
\title{Adds variable(s) to a SpatialPoints}
\usage{
add_variable_spts(sp, df)
}
\arguments{
\item{sp}{A SpatialPoints object.}

\item{df}{A vector or a data frame.}
}
\value{
A SpatialPointsDataFrame
}
\description{
Adds variable(s) to a SpatialPoints
}
\examples{
stations <- imhen::stations
stations <- sf::as_Spatial(stations)
stations <- remove_data_spatialpoints(stations)
stations
add_variable_spts(stations, 1:length(stations))
}
