% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pts_by_poly.R
\name{apply_pts_by_poly}
\alias{apply_pts_by_poly}
\title{Apply functions over points by polygons}
\usage{
apply_pts_by_poly(points, polygons, var, f, ...)
}
\arguments{
\item{points}{a `SpatialPointsDataFrame` object.}

\item{polygons}{a `SpatialPolygonsDataFrame` object.}
}
\description{
This function is a wrapper aroud the `sptools::points_in_polygon` and applies
a function `f` to the variable `var` of the `spatialPointsDataFrame` `points`
object, by the polygons of `spatialPolygonsDataFrame` `polygons` object.
}
\author{
Marc Choisy
}
