% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{areas}
\alias{areas}
\title{Areas of the Polygons of a SpatialPolygon* Object}
\usage{
areas(sppoly)
}
\arguments{
\item{sppoly}{\code{SpatialPolygon*} object as defined in package \code{sp}.}
}
\value{
A list of same length as the input \code{SpatialPolygon*}
object. Each slot of the list is a numeric vector, the lenght of which is
equal to the number of \code{Polygon} objects constituing the polygon entity.
Each of these values are the areas of the of the \code{Polygon} objects,
expressed in the unit of the CRS.
}
\description{
Returns a list of the areas of all the \code{Polygon} objects of the polygons
slot of a \code{SpatialPolygon*} object.
}
\examples{
# working on the provinces polygons of Vietnam:
vn_prov <- gadmVN::gadm()
poly_areas <- areas(vn_prov)
# we can see that some of the polygons are made of several polygons:
sapply(poly_areas, length)
# we can verify that it sums well for the whole country
rgeos::gArea(vn_prov)
sum(sapply(poly_areas, sum))
# as well as for each province:
cbind(rgeos::gArea(vn_prov, TRUE), sapply(poly_areas, sum))

}
