% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_plots.R
\name{barplots_}
\alias{barplots_}
\alias{gg_barplot}
\title{barplots_}
\usage{
barplots_(Spectra, gr = "HistGr", ID = "ID", alpha = 1)

gg_barplot(Spectra, gr = "HistGr", alpha = 1)
}
\arguments{
\item{alpha}{}
}
\examples{

}
