% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_data.R
\name{change_data}
\alias{change_data}
\title{Change the data of a Spatial*DataFrame object.}
\usage{
change_data(spdf, df)
}
\arguments{
\item{spdf}{A Spatial*DataFrame object.}

\item{df}{A data frame.}
}
\value{
An object of the same class as \code{spdf}.
}
\description{
The aim of this function is to propose an alternative to the bug reported
\href{https://github.com/edzer/sp/issues/56}{here}.
}
\examples{
# the working direction
vn <- sptools::gadm("vietnam", "sp", 1, path = getwd(), intlib = TRUE)
vn
change_data(vn, data.frame(gaussian = rnorm(63), uniform = runif(63)))

}
