% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spToolsVectors.R
\name{ci.cor.dep.vec}
\alias{ci.cor.dep.vec}
\title{Confidence Interval for Dependent Pearson Correlation Difference}
\usage{
ci.cor.dep.vec(alpha, cor, cor12, n)
}
\arguments{
\item{alpha}{Alpha level for 1-alpha confidence.}

\item{cor}{Numeric vector of length 2: correlations for measures 1 and 2.}

\item{cor12}{Numeric scalar: correlation between the two predictor variables
associated with \code{cor[1]} and \code{cor[2]}.}

\item{n}{Numeric scalar or numeric vector of length 2 with equal values:
common sample size.}
}
\value{
A 1-row matrix identical to the output of \code{ci.cor.dep}.
}
\description{
A wrapper function for \code{ci.cor.dep} that accepts a vector for the two
correlations and a scalar or repeated vector for the common sample size.
}
\examples{
ci.cor.dep.vec(.05, cor = c(.396, .179), cor12 = .088, n = 166)

}