% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spToolsVectors.R
\name{ci.cor2.vec}
\alias{ci.cor2.vec}
\title{Confidence Interval for Independent Groups Pearson Correlation Difference}
\usage{
ci.cor2.vec(alpha, cor, n)
}
\arguments{
\item{alpha}{Alpha level for 1-alpha confidence.}

\item{cor}{Numeric vector of length 2: correlations for groups 1 and 2.}

\item{n}{Numeric vector of length 2: sample sizes for groups 1 and 2.}
}
\value{
A 1-row matrix identical to the output of \code{ci.cor2}.
}
\description{
A wrapper function for \code{ci.cor2} that accepts vectors for group
correlations and sample sizes.
}
\examples{
ci.cor2.vec(.05, cor = c(.64, .31), n = c(200, 200))

}
