% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spToolsVectors.R
\name{ci.mean.ps.vec}
\alias{ci.mean.ps.vec}
\title{Confidence Interval for Paired Samples Mean Difference}
\usage{
ci.mean.ps.vec(alpha, m, sd, cor, n)
}
\arguments{
\item{alpha}{Alpha level for 1-alpha confidence}

\item{m}{Numeric vector of length 2: means for the two measurements}

\item{sd}{Numeric vector of length 2: standard deviations for the two measurements}

\item{cor}{Estimated correlation between the two measurements}

\item{n}{Sample size (scalar)}
}
\value{
A 1-row matrix identical to the output of ci.mean.ps
}
\description{
A wrapper function for ci.mean.ps that accepts vectors for means and standard deviations.
}
\examples{
ci.mean.ps.vec(.05, c(58.2, 51.4), c(7.43, 8.92), .537, 30)

}
