% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spToolsVectors.R
\name{ci.prop.vec}
\alias{ci.prop.vec}
\title{Confidence Intervals for a Set of One-Sample Proportions}
\usage{
ci.prop.vec(alpha, f, n)
}
\arguments{
\item{alpha}{Alpha level for 1-alpha confidence.}

\item{f}{Numeric vector of frequencies.}

\item{n}{Numeric vector of sample sizes.}
}
\value{
A matrix with one row per element of \code{f} and columns from \code{ci.prop}:
\code{Estimate}, \code{SE}, \code{LL}, and \code{UL}.
}
\description{
A wrapper function for \code{ci.prop} that accepts vectors of frequencies
and sample sizes.
}
\examples{
ci.prop.vec(.05, f = c(120, 95), n = c(300, 250))

}
