% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spToolsVectors.R
\name{ci.prop2.vec}
\alias{ci.prop2.vec}
\title{Confidence Interval for Independent Groups Proportion Difference}
\usage{
ci.prop2.vec(alpha, f, n)
}
\arguments{
\item{alpha}{Alpha level for 1-alpha confidence.}

\item{f}{Numeric vector of length 2: frequencies for groups 1 and 2.}

\item{n}{Numeric vector of length 2: sample sizes for groups 1 and 2.}
}
\value{
A 1-row matrix identical to the output of \code{ci.prop2}.
}
\description{
A wrapper function for \code{ci.prop2} that accepts vectors for frequencies
and sample sizes in two groups.
}
\examples{
ci.prop2.vec(.05, f = c(57, 15), n = c(100, 100))

}
