% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spToolsVectors.R
\name{ci.stdmean.ps.vec}
\alias{ci.stdmean.ps.vec}
\title{Confidence Interval for a Paired Samples Standardized Mean Difference}
\usage{
ci.stdmean.ps.vec(alpha, m, sd, cor, n)
}
\arguments{
\item{alpha}{Alpha level for 1-alpha confidence}

\item{m}{Numeric vector of length 2: means for the two measurements}

\item{sd}{Numeric vector of length 2: standard deviations for the two measurements}

\item{cor}{Estimated correlation between the two measurements (scalar)}

\item{n}{Sample size (scalar)}
}
\value{
A 3-row matrix identical to the output of ci.stdmean.ps
}
\description{
A wrapper function for ci.stdmean.ps that accepts vectors for means and standard deviations.
}
\examples{
ci.stdmean.ps.vec(.05, c(110.4, 102.1), c(15.3, 14.6), .75, 25)

}
