% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spToolsVectors.R
\name{ci.stdmean2.vec}
\alias{ci.stdmean2.vec}
\title{Confidence Interval for a Independent Groups Standardized Mean Difference}
\usage{
ci.stdmean2.vec(alpha, m, sd, n)
}
\arguments{
\item{alpha}{Alpha level for 1-alpha confidence}

\item{m}{Numeric vector of length 2: means for groups 1 and 2}

\item{sd}{Numeric vector of length 2: standard deviations for groups 1 and 2}

\item{n}{Numeric vector of length 2: sample sizes for groups 1 and 2}
}
\value{
A 4-row matrix identical to the output of ci.stdmean2
}
\description{
A wrapper function for ci.stdmean2 that accepts vectors for means, standard deviations, and sample sizes.
}
\examples{
ci.stdmean2.vec(.05, c(35.1, 26.7), c(7.32, 6.98), c(30, 30))

}
