% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup}
\alias{cleanup}
\title{cleanup}
\usage{
cleanup(pattern = "^t")
}
\arguments{
\item{pattern}{A character vector containing the RegEx pattern of the object names to remove. Make sure to surround names by "^...$" if you want to match the name exactly.}
}
\description{
Function which cleans up the GlobalEnv by removing all objects whose name commences with "t."
}
\examples{
Instead of writing:
rm(list=ls(pattern="^t."))
one can write
cleanup()
}
\keyword{rm}
