% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBox.R
\name{createBox}
\alias{createBox}
\title{createBox}
\usage{
createBox(coords_list)
}
\arguments{
\item{coords_list}{The coordinates of the two corners of the box as a list of x and y, as obtained for example with locator().}
}
\description{
This function creates a rectangular SpatialPolygons, given the coordinates of the lower-left and upper-right corners.
}
\examples{
library("sp", "sptools")
createBox(list(x=c(0,2), y=c(0,2)))
}

