% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCircle.R
\name{createCircle}
\alias{createCircle}
\title{createCircle}
\usage{
createCircle(x,y,r,start=0,end=2*pi,nsteps=20)
}
\arguments{
\item{x, y}{The coordinates of the center of the circle.}

\item{r}{The radius of the circle.}

\item{start}{Angle of the start points in radiants.}

\item{end}{Angle of the end points in radiants.}

\item{nsteps}{Number of edges of the polygon approximating the circle.}
}
\description{
This function creates a an approximation of a circle as a SpatialPolygons object.
}
\examples{
# Plot a circle with origin 0,0 and radius 1
library("sp", "sptools")
plot(createCircle(0,0,1))
}

