% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_sppoly.R
\name{cut_poly}
\alias{cut_poly}
\title{Cuts a Polygon object into 2 lines of which it returns the coordinates}
\usage{
cut_poly(poly, pt1, pt2)
}
\arguments{
\item{poly}{A Polygon object as defined the sp package.}

\item{pt1}{A single-point SpatialPoint object as defined in the sp package.}

\item{pt2}{A single-point SpatialPoint object as defined in the sp package.}
}
\value{
a list of 2 2-variable data frames of coordinates. In each data frame
the first column is the longitude and the second column is the latitude.
}
\description{
The polygon is cut between the points of the polygon that are closest to pt1
and pt2
}
