% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_plots.R
\name{plot_cutoffs}
\alias{plot_cutoffs}
\alias{cutoffs_plot}
\alias{cutoffs_plot}
\alias{cutoffs_calcultate_stats}
\alias{cutoffs_plot}
\alias{cutoffs_get_df}
\alias{cutoffs_plot}
\alias{cutoffs_plot_df}
\title{sp_plot_range_and_cutoffs}
\usage{
plot_cutoffs(...)

cutoffs_plot(obj, percent = 95, return_as = c("plotly", "ggplot", "list"))

cutoffs_calcultate_stats(sp, percent = 100, center = FALSE, scale = FALSE)

cutoffs_get_df(data, cutoffs, percent = 100, scaling = FALSE,
  centering = scaling, percent.cutoffs = 95)

cutoffs_plot_df(df, lwd_mean = 1, alpha = 0.3, COLORS = c("#e41a1c",
  "#377eb8", "black"))
}
\arguments{
\item{obj}{a list of hyperSpec objects}

\item{percent}{percentage of y axis range to be plotted}

\item{df}{a data frame}

\item{lwd_mean}{width of line for mean}

\item{alpha}{transparency}

\item{COLORS}{vector with either names or codes of 3 colors}

\item{size}{size of line for means}

\item{limsY}{limits for y axis}
}
\description{
sp_plot_range_and_cutoffs
}
\examples{
cutoffs_plot(obj)

}
