% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_bbox_proj.R
\name{define_bbox_proj}
\alias{define_bbox_proj}
\title{Defines new boundaries box and projections of a sf object}
\usage{
define_bbox_proj(sf_obj, boundbox, crs)
}
\arguments{
\item{sf_obj}{an objet of class "sf".}

\item{boundbox}{character, bounding box.}

\item{crs}{character, coordinate reference system.}
}
\description{
The function defines the attributes \code{bbox} and \code{crs} of a sf
object.
}
\examples{
library(sf)
vn <- gadm("Vietnam", "sf", 0)
vn
bb <- st_bbox(vn)
crs <- st_crs(vn)

vn2 <- gadmVN::gadm()
vn2 <- define_bbox_proj(vn2, bb, crs)
vn2

}
