% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_sppoly.R
\name{df2splist}
\alias{df2splist}
\title{Transform a 2-variable dataframe into a list of single-point SpatialPoints.}
\usage{
df2splist(df)
}
\arguments{
\item{df}{a 2-variable dataframe}
}
\value{
a list of single-point SpatialPoints
}
\description{
Transform a 2-variable dataframe into a list of single-point SpatialPoints.
}
