% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_sppoly.R
\name{dist_pt_poly}
\alias{dist_pt_poly}
\title{Distance from a point to all the points of a data frame of coordinates.}
\usage{
dist_pt_poly(pt, poly_coord_df)
}
\arguments{
\item{pt}{a SpatialPoints* object}

\item{poly_coord_df}{a data frame of coordinates with first column being the
longitude and the second column being the latitude.}
}
\description{
Distance from a point to all the points of a data frame of coordinates.
}
