% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ROC_CV_by_gr.R
\name{do_ROC_CV_by_gr_2}
\alias{do_ROC_CV_by_gr_2}
\title{do_ROC_CV_by_gr}
\usage{
do_ROC_CV_by_gr_2(Spectra, variables_to_analyze = "CitoGr", k_folds = 3,
  times = 10, parallelization = FALSE)
}
\description{
do_ROC_CV_by_gr
}
\examples{
do_ROC_CV_by_gr_2(Spectra, "gr")
}
