% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_HTML.R
\name{Fig}
\alias{Fig}
\alias{Fig_1col}
\alias{Fig_2cols}
\alias{Fig_3cols}
\alias{insert_fig_2cols}
\alias{insert_fig_3cols}
\title{Fig}
\usage{
Fig(file = "", caption = "", fig.align = c("center", "left", "right",
  "justify")[1], width = 500, text.align = c("center", "justify", "left",
  "right")[1], text.style = c("italic", "normal", "oblique")[1],
  class = "figure", overwrite = FALSE, quality = 85)

Fig_1col(file = "", caption = "", width = 300, fig.align = c("center",
  "left", "right", "justify")[1], text.align = c("justify", "center", "left",
  "right")[1], text.style = c("italic", "normal", "oblique")[1],
  class = "col-sm-12", overwrite = FALSE, quality = 85)

Fig_2cols(file = "", caption = "", width = 300, fig.align = c("center",
  "left", "right", "justify")[1], text.align = c("justify", "center", "left",
  "right")[1], text.style = c("italic", "normal", "oblique")[1],
  class = "col-sm-6", overwrite = FALSE, quality = 85)

Fig_3cols(file = "", caption = "", width = 300, fig.align = c("center",
  "left", "right", "justify")[1], text.align = c("justify", "center", "left",
  "right")[1], text.style = c("italic", "normal", "oblique")[1],
  class = "col-sm-4", overwrite = FALSE, quality = 85)

insert_fig_2cols(file = "", caption = "", width = 300,
  fig.align = c("center", "left", "right", "justify")[1],
  text.align = c("justify", "center", "left", "right")[1],
  text.style = c("italic", "normal", "oblique")[1], class = "col-sm-6",
  overwrite = FALSE, quality = 85)

insert_fig_3cols(file = "", caption = "", width = 300,
  fig.align = c("center", "left", "right", "justify")[1],
  text.align = c("justify", "center", "left", "right")[1],
  text.style = c("italic", "normal", "oblique")[1], class = "col-sm-4",
  overwrite = FALSE, quality = 85)
}
\arguments{
\item{quality}{}
}
\description{
Fig
}
