% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_summary.R
\name{ggdensity}
\alias{ggdensity}
\alias{ggdensity}
\alias{ggdensity_3subplots}
\alias{ggdensity}
\alias{ggdensity_age}
\title{ggdensity}
\usage{
ggdensity(sp, x = "age", gr = NULL)

ggdensity_3subplots(sp, x = "age", gr = c("Hybrid_normal",
  "Hybrid_cervicitis", "Hybrid_cin3_cis"))

ggdensity_age(sp, x = "age", gr = "Hybrid_normal")
}
\arguments{
\item{gr}{}
}
