% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_plots.R
\name{ggplot_mean_and_range}
\alias{ggplot_mean_and_range}
\title{ggplot_mean_and_range}
\usage{
ggplot_mean_and_range(x, by = "Compared", percent = 95, limsY = c(0, 1))
}
\arguments{
\item{percent}{}
}
\description{
ggplot_mean_and_range
}
