% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfixPaste.R
\name{\%_\%}
\alias{\%_\%}
\title{Paste without separator - infix operator}
\usage{
x \%_\% y
}
\arguments{
\item{x}{An object of type character.}

\item{y}{An object of type character.}
}
\description{
Infix operator to combine strings. Equivalent to paste(..., sep="").
}
\examples{
z<-5
"The result is: " \%_\% z \%_\% "."
> The result is 5.
}
\keyword{paste}
