% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid2raster.R
\name{grid2raster}
\alias{grid2raster}
\title{Convert an SpatialPoints* grid into a Raster*}
\usage{
grid2raster(grid)
}
\arguments{
\item{grid}{A SpatialPointsDataFrame grid (i.e. where points are regularly
spaced).}
}
\value{
A Raster* object.
}
\description{
Convert an SpatialPoints* grid into a Raster*
}
\examples{
library(sf)
vn <- gadmVN::gadm(level = "country")
vn <- sf::as_Spatial(vn)
grid <- make_grid(vn, 100)
grid
grid2raster(grid)

# An example with data:
grid2 <- add_variable_spts(grid, 1:length(grid))
grid2raster(grid2)


}
