% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head1.R
\name{head1}
\alias{head1}
\title{head1}
\usage{
head1(x, n = 1)
}
\arguments{
\item{x}{A vector, matrix, table, data frame or function.}

\item{n}{The number of rows to return. Defaults to 1.}
}
\description{
Function to replicate the head() function from the 'utils' package, but with a default of returning only one row.
}
\examples{
# Return one row
head1(mtcars)
# Return multiple rows
head1(x = mtcars, n = 3)

}
\keyword{head}
