% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{i}
\alias{i}
\title{i}
\usage{
i(x, by = 1)
}
\arguments{
\item{x}{A numeric.}

\item{by}{The value to increment x by. Defaults to 1.}
}
\description{
Function to increment a numeric.
}
\examples{
Increments. For example:
x<-5
i(x)
After this operation, x == 6.
}
\keyword{increment}
