% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largest_polygons.R
\name{largest_polygons}
\alias{largest_polygons}
\title{Filter the Largest Polygons of each polygon slot of a SpatialPolygon* object}
\usage{
largest_polygons(sppoly)
}
\arguments{
\item{sppoly}{\code{SpatialPolygon*} object as defined in package \code{sp}.}
}
\value{
An object of the same class as \code{sppoly} where each slot of the
\code{polygons} element is made on one single \code{Polygon} object that is
the largest \code{Polygon} is the input \code{sppoly}.
}
\description{
Filter the Largest Polygons of each polygon slot of a SpatialPolygon* object
}
\examples{
vn_prov <- gadmVN::gadm()
vn_prov2 <- largest_polygons(vn_prov)
sp::plot(vn_prov)
sp::plot(vn_prov2)
# same with the country level:
vn <- gadmVN::gadm(level = "country")
vn2 <- largest_polygons(vn)
rgeos::gEnvelope(vn)

}
