% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linedf2SpatialLines.R
\name{linedf2SpatialLines}
\alias{linedf2SpatialLines}
\title{linedf2SpatialLines}
\usage{
linedf2SpatialLines(linedf)
}
\arguments{
\item{linedf}{a data frame with the following columns:
id = generic ids of the lines,
fx = coordinates x of the first point of the line,
fy = coordinates y of the first point of the line,
tx = coordinates x of the second point of the line,
tx = coordinates y of the second point of the line.}
}
\description{
This function converts a dataframe of IDs and coordinates into a SpatialLines object of the package "sp".
}
\examples{
library("sp", "sptools")
sl = linedf2SpatialLines(data.frame(id=1, fx=0, fy=0, tx=2, ty=2))
class(sl)
plot(sl)
}

