% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Sample Point Locations in a SpatialPolygon*}
\usage{
make_grid(sppoly, n, ...)
}
\arguments{
\item{sppoly}{\code{SpatialPolygon*} object as defined in the \code{sp} package.}

\item{n}{approximate sample size.}

\item{...}{optional arguments passed to \code{sp::makegrid}}
}
\value{
A \code{SpatialPoins$} object. The length of the object is
approximately equal to \code{n}.
}
\description{
This is a wrapper around the \code{sp::makegrid} function, only for
\code{SpatialPolygon*} objects.
}
\examples{
library(sf)
vn <- gadmVN::gadm(level = "country")
vn <- sf::as_Spatial(vn)
sp::plot(vn)
grid <- make_grid(vn, 100)
length(grid)
sp::plot(grid, add = TRUE)
grid2 <- make_grid(vn, cellsize = .5)
length(grid2)
sp::plot(grid2, add = TRUE, col = "blue")

}
