% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_hyperSpec.R
\name{make_info}
\alias{make_info}
\title{Put `Object`, `Laser_power` and `Integration_time` into one variable}
\usage{
make_info(sp)
}
\description{
Put `Object`, `Laser_power` and `Integration_time` into one variable
}
