% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_hyperSpec.R
\name{merge_sp_and_diagnosis}
\alias{merge_sp_and_diagnosis}
\title{merge_sp_and_diagnosis}
\usage{
merge_sp_and_diagnosis(SP, add.data, by.y = "ID2", by.x = "Object")
}
\arguments{
\item{by.x}{}
}
\description{
merge_sp_and_diagnosis
}
