% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multdiff.R
\name{multdiff}
\alias{multdiff}
\title{multdiff}
\usage{
multdiff(x)
}
\arguments{
\item{x}{A list of vectors.}
}
\description{
Performs setdiff on the elements of x, where x is a list of vectors of the same type but not necessarily the same length.
}
\examples{
Instead of writing, for example:
setdiff(setdiff(a,b),c)
one can write
multdiff(list(a,b,c))
}
\keyword{setdiff}
