% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSetFun.R
\name{MultiSetFun}
\alias{MultiSetFun}
\title{MultiSetFun}
\usage{
MultiSetFun(fun, x)
}
\arguments{
\item{fun}{A set function, i.e., one out of intersect, setdiff, setequal, union.}

\item{x}{A list of vectors.}
}
\description{
Performs a set function (e.g. intersect, setdiff) on x, where x is a list of vectors of the same type but not necessarily the same length. Taken directly from W. Lee Pang at https://gist.github.com/wleepang/4724679.
}
\examples{
Instead of writing, for example:
setdiff(setdiff(a,b),c)
one can write
MultiSetFun(setdiff,list(a,b,c))
}
\keyword{functions}
\keyword{set}
