% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_exclude.R
\name{na_exclude}
\alias{na_exclude}
\title{Exclude features with NA attributes}
\usage{
na_exclude(spdf)
}
\arguments{
\item{spdf}{A \code{SpatialPointsDataFrame} object.}
}
\value{
A \code{SpatialPointsDataFrame} object.
}
\description{
This function removes the features of a \code{SpatialPointsDataFrame} for
which there is NA is the attributes.
}
\examples{
library(imhen)

stations <-  imhen::stations
stations <- sf::as_Spatial(stations)
stations@data$elevation[seq(2, 12, 2)] <- NA

head(stations@data)
head(na_exclude(stations)@data)
}
