% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Image.R
\name{plot.Image}
\alias{plot.Image}
\title{Plot image of class `EBImabe`}
\usage{
plot.Image(x, title = deparse(substitute(x), width.cutoff = 500L, nlines = 1),
  method = "raster", all = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot image of class `EBImabe`
}
