% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sp_summ_stat.R
\name{plot_sp_summ_stat}
\alias{plot_sp_summ_stat}
\title{Plot many summary plots of spectra}
\usage{
plot_sp_summ_stat(sp, vars, zones)
}
\arguments{
\item{zones}{}
}
\description{
(*** NOT RELEVANT ***)
}
\details{
Plot many summary plots of spectra (plot_sp_summ_stat)
}
\examples{

}
