% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_summary.R
\name{print_gr_summary}
\alias{print_gr_summary}
\title{Print summary of indicated variable}
\usage{
print_gr_summary(sp, ID = "ID", gr = "gr")
}
\arguments{
\item{ID}{variable with IDs of medical samples.}

\item{gr}{name of factor variable.}
}
\description{
Print summary of indicated variable
}
