% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_precisely.R
\name{print_precisely}
\alias{print_precisely}
\title{print_precisely}
\usage{
print_precisely(x)
}
\arguments{
\item{x}{A numeric, character or factor variable that can be transformed to numeric.}
}
\description{
Function to transform the input to numeric and then print it with maxiumum precision.
}
\examples{
print_precisely(x)
}
\keyword{numeric,}
\keyword{precision}
