% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projections}
\alias{projections}
\alias{proj0}
\alias{projVN}
\title{Projections}
\format{CRS objects as defined in the \code{sp} package:
\describe{
  \item{proj0}{unprojected coordinates on the GWS84 datum}
  \item{projVN}{EPSG:3405 aka VN-2000 aka UTM48N}
}}
\source{
\url{http://spatialreference.org}
}
\usage{
proj0

projVN
}
\description{
Pre-defined CRS
}
\examples{
library(sf)
# working on the provinces polygons of Vietnam:
vn_prov <- gadmVN::gadm()
vn_prov <- sf::as_Spatial(vn_prov)
crs(vn_prov)
vn_prov2 <- sp::spTransform(vn_prov, projVN)
crs(vn_prov2)
vn_prov3 <- sp::spTransform(vn_prov2, crs(vn_prov))
crs(vn_prov)
}
\keyword{datasets}
