% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_hyperSpec.R
\name{read_spectra}
\alias{read_spectra}
\alias{read_spectra_ts}
\title{read_spectra}
\usage{
read_spectra(path, pattern = ".*", read_fun = c("auto", "ascii", "ts",
  "a FUNCTION")[1], add.data = NULL, by.x = "Object", by.y = "ID2", ...)

read_spectra_ts(path, pattern = ".*", read_fun = "ts", add.data = NULL,
  by.x = "Object", by.y = "ID2", ...)
}
\arguments{
\item{...}{}
}
\description{
read_spectra
}
\details{
\code{read_spectra_ts} is a wrapper of \code{read_spectra} for
time series.
}
