% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_data.R
\name{remove_data_spatialpoints}
\alias{remove_data_spatialpoints}
\title{Transform SpatialPointsDataFrame into SpatialPoints}
\usage{
remove_data_spatialpoints(spdf)
}
\arguments{
\item{spdf}{A SpatialPointsDataFrame object.}
}
\value{
A SpatialPoints object.
}
\description{
Transform SpatialPointsDataFrame into SpatialPoints
}
