% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_data.R
\name{remove_data_spatialpolygons}
\alias{remove_data_spatialpolygons}
\title{Transform SpatialPolygonsDataFrame into SpatialPolygons}
\usage{
remove_data_spatialpolygons(spdf)
}
\arguments{
\item{spdf}{A SpatialPolygonDataFrame object.}
}
\value{
A SpatialPolygon object.
}
\description{
Transform SpatialPolygonsDataFrame into SpatialPolygons
}
\examples{
vn <- gadm("vietnam", "sp", 1)
vn
remove_data_spatialpolygons(vn)

}
