% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_grid.R
\name{res_grid}
\alias{res_grid}
\title{Resolution of a grid}
\usage{
res_grid(grdsppts)
}
\arguments{
\item{grdsppts}{a grid \code{SpatialPoints} object as returned by
\code{sp::makegrid} or \code{sptools::make_grid}.}
}
\description{
Equivalent to the `raster::res` function but for a grid `SpatialPoints`
object as returned by `sp::makegrid` or `sptools::make_grid`.
}
\examples{
library(sf)
vn <- gadmVN::gadm(level = "country")
vn <- sf::as_Spatial(vn)
sp::plot(vn)
grid <- make_grid(vn, 100)
res_grid(grid)
}
\author{
Marc Choisy
}
