% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_from_grid.R
\name{resample_from_grid}
\alias{resample_from_grid}
\title{Resample a Raster object from a Grid}
\usage{
resample_from_grid(rstr, grd)
}
\arguments{
\item{rstr}{A \code{Raster*} object (i.e. a \code{RasterLayer} or a
\code{RasterBrick} object).}

\item{grd}{A \code{SpatialPoint} object with points regularly spaced.}
}
\value{
A \code{RasterLayer} object.
}
\description{
This function is a wrapper around the \code{\link[raster]{resample}}
function. Instead of resampling a Raster object from another Raster object as
in \code{\link[raster]{resample}}, this function resamples a Raster object
from a \code{SpatialPoints*} grid object. In case of a \code{RasterBrick} or
a \code{RasterStack} object, it resamples only the first layer and returns it
as a \code{RasterLayer} object.
}
\examples{
\dontrun{
library(sp)
library(raster)
library(wordlpopVN)


# # download vietnam country administrative map in the internal library and in
# the working direction
vn <- sptools::gadm("vietnam", "sf", 0, intlib = TRUE, save = TRUE)
# The raster of population
ppp2010 <- worldpopVN::getpop(2010)

# A grid of 100 points over the country:
proj <- proj4string(country)
grid100 <- makegrid(country, 100)
grid100 <- SpatialPoints(grid100, CRS(proj))

# Let's resample:
ppp2010rspld <- resample_from_grid(ppp2010, grid100)

# Let's compare:
ppp2010
ppp2010rspld

plot(ppp2010rspld)
plot(grid100, add = TRUE)
plot(country, add = TRUE)
}
}
