% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_raster.R
\name{rescale_raster}
\alias{rescale_raster}
\title{Rescales the values of a raster as weights}
\usage{
rescale_raster(rstr)
}
\arguments{
\item{rstr}{a RasterLayer object}
}
\value{
A RasterLayer object with values between 0 and 1 and summing to 1.
}
\description{
This function rescales the values of a raster between 0 and 1 so that they
sum to 1. These values can then be used as weights.
}
\examples{

library(worldpopVN)
provinces <- sptools::gadm("vietnam", "sp", 1, intlib = FALSE, save = TRUE)
ppp2010 <- worldpopVN::getpop(2010)

hanoi <- sptools::crop_on_poly(ppp2010, subset(provinces,
  VARNAME_1 == "Ha Noi"))

# before rescaling:
sum(values(hanoi), na.rm = TRUE)

# after rescaling:
sum(values(rescale_raster(hanoi)), na.rm = TRUE)

}
