% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_largest_polygons.R
\name{rm_largest_polygons}
\alias{rm_largest_polygons}
\title{Filter out the Largest Polygon of each polygon slot of a SpatialPolygon* object}
\usage{
rm_largest_polygons(sppoly, subset)
}
\arguments{
\item{sppoly}{\code{SpatialPolygon*} object as defined in package \code{sp}.}

\item{subset}{a subsetting condition}
}
\value{
An object of the same class as \code{sppoly} where, for each slot of
the \code{polygons} element that originally contains more than one polygon,
the largest one has been removed.
}
\description{
Filter out the Largest Polygon of each polygon slot of a SpatialPolygon* object
}
\examples{
vn_prov <- gadmVN::gadm()
vn_prov2 <- rm_largest_polygons(vn_prov)
sp::plot(vn_prov)
sp::plot(vn_prov2)
# same with the country level:
vn <- gadmVN::gadm(level = "country")
vn2 <- rm_largest_polygons(vn)
rgeos::gEnvelope(vn)
# extracting the largest polygon only for Da Nang:
vn_prov3 <- rm_largest_polygons(vn_prov, province == "Da Nang")
sp::plot(vn_prov3)
# extracting the largest polygon only for Da Nang and Khanh Hoa:
vn_prov4 <- rm_largest_polygons(vn_prov, province \%in\% c("Da Nang", "Khanh Hoa"))
sp::plot(vn_prov4)

}
\seealso{
\code{\link[sptools]{largest_polygons}}
}
