% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_with_cv_3.R
\name{roc_manyroc_cv_by_variable}
\alias{roc_manyroc_cv_by_variable}
\title{sp_manyroc_cv_by_variable}
\usage{
roc_manyroc_cv_by_variable(Spectra, variables_to_analyze, k_folds = 3,
  times = 10)
}
\description{
sp_manyroc_cv_by_variable
}
\examples{
sp_manyroc_cv_by_variable(Spectra2, "gr")
}
