% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_classification.R
\name{ROC_with_cv}
\alias{ROC_with_cv}
\alias{plot_10}
\title{ROC_with_cv}
\usage{
ROC_with_cv(Spectra_gr, Var, k = 5, times = 3, seeds = 2, n_min = k,
  range_percent = 95)

plot_10(Spectra_gr, Var, k = 5, times = 3, seeds = 2, n_min = k,
  range_percent = 95)
}
\arguments{
\item{n_min}{}
}
\description{
ROC_with_cv
}
\examples{
Spectra_gr <- Spectra
Var <- "CitoGr"

}
