% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_classification.R
\name{ROC_with_cv_2}
\alias{ROC_with_cv_2}
\title{ROC_with_cv_2}
\usage{
ROC_with_cv_2(Spectra_gr, Var, k = 5, times = 10, seeds = 2, n_min = k,
  range_percent = 95)
}
\arguments{
\item{n_min}{}
}
\description{
ROC_with_cv_2
}
\examples{
Spectra_gr <- Spectra
Var <- "CitoGr"
}
