% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample1.R
\name{sample1}
\alias{sample1}
\title{sample1}
\usage{
sample1(from = 1, to = 100, by = 1, size = 1)
}
\arguments{
\item{from}{Starting value of the range ob possible values.}

\item{to}{Ending value of the range of possible values.}

\item{by}{Step size for the range of possible values.}

\item{size}{Number of random values to draw.}
}
\description{
Returns one (default) or several numbers from a range of numbers generated by the seq() function.
}
\examples{
# Return one integer out of the set {1, 2,..., 100}
sample1()
# Return two integers out of the set {0, 2.5, 5, 7.5, 10}
sample1(from = 0, to = 10, by = 2.5, size = 2)

}
\keyword{sample}
