% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_aggregate_lst.R
\name{sf_aggregate_lst}
\alias{sf_aggregate_lst}
\title{Aggregates sf from a list of event}
\usage{
sf_aggregate_lst(df_sf, history_lst, from, to = "2018-12-31")
}
\arguments{
\item{df_sf}{A sf data frame containing at least the variables
\code{province}, \code{geometry} and \code{district} if `history_lst`
contains complex event}

\item{history_lst}{A list containing a list of event, each code with a slot
\code{after}, a slot \code{before}, a slot{event} (split/merge/rename/
complex merge/complex split) and a slot \code{year}.}

\item{from}{Initial date of the time range selected for the province
definition, of the class \code{Date}, \code{character} or \code{numeric}.}

\item{to}{Final date of the time range selected for the province
definition, of the class \code{Date}, \code{character} or \code{numeric}, by
default  \code{"2018-12-31"}}
}
\value{
A object of the same class as \code{df_sf} in which all the variables
and geometry that needed to be aggregated or renamed (according to the time
range) are changed.
}
\description{
Tidy the data and merges data accordingly to a time range and by the
values concerned by a split/merge/rename event and return a sf data frame
for the time range imputed.
}
\details{
For each `split event`, the geometry of the variables contained  in the slot
`after` is combined and rename by the variable in the slot `before`.
 For each `rename event`, the variable contained in the slot `after` is
 renamed.
 In the new rows, <NA> will be added in the other column than `sel` and
 `geometry`
}
\examples{
 # to have the list of split/merge/rename event for Vietnam
 vn_history <- dictionary::vn_history

 vn_prov04 <- gadmVN::gadm(date = "2004-01-01")
 vn_prov70 <- sf_aggregate_lst(vn_prov04, vn_history, from = "1970",
                               to = "2004")

}
