% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfpts2df.R
\name{sfpts2df}
\alias{sfpts2df}
\title{Transforms an sf points into a dataframe}
\usage{
sfpts2df(sfpts)
}
\arguments{
\item{sfpts}{An \code{sf} points object.}
}
\value{
A data frame that contains the same fields as \code{sfpts} plus the
longitude and latitude in 2 separate columns.
}
\description{
Transforms an sf points into a dataframe
}
\examples{
sfpts2df(worldpopVN::popcenters2009)

}
