% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spToolsOutput.R
\name{sp.infer}
\alias{sp.infer}
\title{Add test statistic, df, and p-value to CI output}
\usage{
sp.infer(results, df = NULL, n = NULL, null_value = 0, conf_level = 0.95)
}
\arguments{
\item{results}{1-row matrix output from Bonett CI functions (with Estimate, SE, LL, UL)}

\item{df}{Optional degrees of freedom (if NULL, try to infer from n)}

\item{n}{Optional sample size to infer df if df is missing}

\item{null_value}{Hypothesized null value for test statistic (default 0)}
}
\value{
1-row matrix with columns:
Estimate, SE, t, df, p, LL, UL
}
\description{
Add test statistic, df, and p-value to CI output
}
