% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spToolsObject.R
\name{sp.rows}
\alias{sp.rows}
\title{Filter Rows from Data Frame, Matrix, or List}
\usage{
sp.rows(out, rows = NULL)
}
\arguments{
\item{out}{A data frame, matrix, atomic vector, or a list of such objects.}

\item{rows}{A numeric vector of row indices or a character vector of row names to retain. If \code{NULL}, all rows are returned.}
}
\value{
A filtered object with only the specified rows. The output will match the input type (matrix, data frame, or list).
}
\description{
Filters specified rows from a data frame, matrix, or list of such objects.
}
\examples{
df <- data.frame(A = 1:5, B = letters[1:5])
sp.rows(df, rows = c(1, 3))

}
