% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_manyroc_cv_by_variable.R
\name{sp_manyroc_cv_by_variable}
\alias{sp_manyroc_cv_by_variable}
\alias{sp_roc_with_cv_3}
\alias{print.hide_it}
\title{sp_manyroc_cv_by_variable}
\usage{
sp_manyroc_cv_by_variable(Spectra, variables_to_analyze, k_folds = 3,
  times = 10, seeds = 2222222)

sp_roc_with_cv_3(Var, Spectra_gr, k = 5, times = 10, seeds = 2222222,
  n_min = k)

\method{print}{hide_it}(x, ...)
}
\arguments{
\item{Spectra}{hyperSpec object}

\item{variables_to_analyze}{Character list}

\item{k_folds}{positive integer}

\item{times}{positive integer}

\item{seeds}{NULL | vector of integers}
}
\description{
sp_manyroc_cv_by_variable
}
\examples{
sp_manyroc_cv_by_variable(Spectra2, "gr")
}
