% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLines2df.R
\name{SpatialLines2df}
\alias{SpatialLines2df}
\title{SpatialLines2df}
\usage{
SpatialLines2df(spatial_line)
}
\arguments{
\item{spatial_line}{An object class SpatialLinesDataFrame from the package sp.}
}
\description{
This function extracts the coordinates of the start and end points of each segment of a Spatial Line and returns a dataframe.
}
\examples{
library("sp","sptools")
c1 = cbind(c(0,2), c(0,2))
l1 = Line(c1)
sl = list(Lines(list(l1), ID = 1))
SpatialLines2df(SpatialLines(sl))
}

