% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPClusterSE.R
\name{SPClusterSE}
\alias{SPClusterSE}
\title{SPClusterSE}
\usage{
SPClusterSE(Data, Model, ClusterVar)
}
\arguments{
\item{Data}{A dataframe containing the data the regression is run in.}

\item{Model}{A saved regression model.}

\item{ClusterVar}{The variable to cluster by.}
}
\description{
Function which calculates clustered standard errors for a regression model. This function is basically a copy of Mahmood Ara's work (http://www.ne.su.se/polopoly_fs/1.216115.1426234213!/menu/standard/file/clustering1.pdf), as described by Kevin Goulding (https://thetarzan.wordpress.com/2011/06/11/clusteredstandarderrorsinr/). The example is also adapted from Kevin's excellent post.
}
\examples{
require(foreign)
nmar = read.dta("http://www.montana.edu/econ/cstoddard/562/wr-nevermar.dta")
# Run a plain linear regression
regt = lm(nevermar ~ impdum, data = nmar)
# apply the 'SPClusterSE' function by choosing a variable to cluster on.
# here, we are clustering on state.
SPClusterSE(nmar, regt, nmar$state)
}
\keyword{clustered}
\keyword{errors}
\keyword{standard}
