% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPdhms.R
\name{SPdhms}
\alias{SPdhms}
\title{SPdhms}
\usage{
SPdhms(t)
}
\arguments{
\item{x}{A numeric containing the number of seconds to transform.}
}
\description{
Function to transform seconds into "days hours:minutes:seconds" format. Taken from https://stackoverflow.com/questions/27312292/convert-seconds-to-days-hoursminutesseconds.
}
\examples{
SPdhms(48000)
"0 13:20:00"
}
\keyword{date,}
\keyword{days}
\keyword{hours,}
\keyword{minutes,}
\keyword{seconds,}
\keyword{time,}
\keyword{transformation,}
