% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_spdf.R
\name{split_spdf}
\alias{split_spdf}
\title{Splits a SpatialPolygons* into a list of SpatialPolygons*}
\usage{
split_spdf(spdf)
}
\arguments{
\item{spdf}{a SpatialPolygons* object.}
}
\value{
a list of SpatialPolygons*
}
\description{
Splits a SpatialPolygons* into a list of SpatialPolygons*
}
\author{
Marc Choisy
}
