% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitLines.R
\name{splitLines}
\alias{splitLines}
\title{splitLines}
\usage{
splitLines(spatial_line, split_length = 10, return.dataframe = FALSE,
  plot.results = FALSE)
}
\arguments{
\item{spatial_line}{An "sp" object of class SpatialLines or SpatialLinesDataFrame.}

\item{split_length}{The length of the segments to split the lines into, in units of the SpatialLines. Default 10.}

\item{return.dataframe}{If TRUE, the functions returns a dataframe with IDs and coordinates, otherwise returns a SpatialLines object. Default FALSE.}

\item{plot.results}{Visual representation of the results. If TRUE, the functions plots the original object in black and the vertex of the new segments in red. Default FALSE.}
}
\description{
This function splits each line of an "sp" object of class SpatialLines into segments of a given length.
}
\examples{
library("sp", "sptools")
c1 = cbind(c(0,2,5,7), c(0,2,3,3))
l1 = Line(c1)
sl = list(Lines(list(l1), ID = 1))
SP = SpatialLines(sl)
plot(SP)
SPs = splitLines(SP, 0.4, plot.results = TRUE)
}
\references{
http://math.stackexchange.com/questions/175896/finding-a-point-along-a-line-a-certain-distance-away-from-another-point?newreg=468f66d7274f449b8ecf3fa4e63f41fe

http://tutorial.math.lamar.edu/Classes/CalcII/Vectors_Basics.aspx
}

