% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLoadPackages.R
\name{SPLoadPackages}
\alias{SPLoadPackages}
\title{SPLoadPackages}
\usage{
SPLoadPackages(packages)
}
\arguments{
\item{packages}{A character vector containing the names of packages to load.}
}
\description{
Function which compares two dataframes (the second with only a single row) column-wise, yielding a Boolean vector of equality between the two. Can be used to select a specific row out of the first (Data), which corresponds in some columns to the same columns in the second (ComparisonVector).
}
\examples{
Instead of writing, for example:

if (!require(dplyr)) install.packages('dplyr')
library(dplyr)
if (!require(zoo)) install.packages('zoo')
library(zoo)

one can write

SPLoadPackages(c("dplyr","zoo"))
}
\keyword{packages}
