% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPNum.R
\name{SPNum}
\alias{SPNum}
\title{SPNum}
\usage{
SPNum(x)
}
\arguments{
\item{x}{A numeric, character or factor variable that can be transformed to numeric.}
}
\description{
Function to transform factor to numeric. Returns input argument as transformed to numeric.
}
\examples{
Instead of writing, for example:
as.numeric(as.character(x))
one can write
SPNum(x)
}
\keyword{factor,}
\keyword{numeric}
