% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPPause.R
\name{SPPause}
\alias{SPPause}
\title{SPPause}
\usage{
SPPause()
}
\description{
Function which pauses for a keypress.
This function is derived from https://diego.assencio.com/?index=86c137b502561d44b8be02f06d80ee16, accessed 16.01.2018.
}
\examples{
SPPause()

}
\keyword{keypress}
\keyword{pause,}
