% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPSubset.R
\name{SPSubset}
\alias{SPSubset}
\title{SPSubset}
\usage{
SPSubset(Data, ComparisonVector, by.Data, by.ComparisonVector = by.Data)
}
\arguments{
\item{Data}{A dataframe containing the data you want to select a specific row from.}

\item{ComparisonVector}{A dataframe with a single row which contains the indices used to select a row from dataframe Data.}

\item{by.Data}{A character vector of column names of dataframe Data which shall serve as indices for row selection.}

\item{by.ComparisonVector}{A character vector of column names of dataframe ColumnVector which shall serve as indices for row selection. Needs to have equal length as by.Data and defaults to by.Data.}
}
\description{
Function which compares two dataframes (the second with only a single row) column-wise, yielding a Boolean vector of equality between the two. Can be used to select a specific row out of the first (Data), which corresponds in some columns to the same columns in the second (ComparisonVector).
}
\examples{
Instead of writing, for example:
A[A$a==B$a & A$b==B$b & A$c==B$c & A$d==B$d,]
one can write
A[SPSubset(A,B,c("a","b","c","d")),]
}
\keyword{subset}
