% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptools.R
\docType{package}
\name{sptools}
\alias{sptools}
\alias{sptools-package}
\title{sptools: Additional Functions for Spatial* Objects}
\description{
The package contains a collection of functions to extend the packages
\code{sp} and \code{rgeos}.
}
