% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_interactive_plots.R, R/for_plots.R
\name{sub_plot}
\alias{sub_plot}
\alias{sub_plot_range}
\alias{sub_plot_means}
\title{sub_plot}
\usage{
sub_plot(..., shareX = T, shareY = T, titleX = T, titleY = T,
  title = "")

sub_plot_range(obj)

sub_plot_means(obj)
}
\arguments{
\item{title}{}
}
\description{
sub_plot

sub_plot_range

sub_plot_means
}
\examples{
qplot_spRange(chondro, "clusters") +
    layer_spStat(chondro,"clusters", mean, size = 1)

}
