% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_plots.R
\name{sub_plot_cutfoos}
\alias{sub_plot_cutfoos}
\title{sub_plot_range_and_means}
\usage{
sub_plot_cutfoos(obj, percent = 95, size = 0.8, limsY = c(0, 1),
  return_as = c("plotly", "ggplot"))
}
\arguments{
\item{obj}{a list of hyperSpec objects}

\item{percent}{percentage of y axis range to be plotted}

\item{size}{size of line for means}

\item{limsY}{limits for y axis}
}
\description{
sub_plot_range_and_means
}
