% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_interactive_plots.R
\name{supbplots_4}
\alias{supbplots_4}
\title{4 subplots (plotly)}
\usage{
supbplots_4(sp__i, sp, gr = "HistGr")
}
\arguments{
\item{gr}{}
}
\description{
supbplots_4: 2 types of spectra in 4 (2x2) subplots
}
