% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail1.R
\name{tail1}
\alias{tail1}
\title{tail1}
\usage{
tail1(x, n = 1)
}
\arguments{
\item{x}{A vector, matrix, table, data frame or function.}

\item{n}{The number of rows to return. Defaults to 1.}
}
\description{
Function to replicate the tail() function from the 'utils' package, but with a default of returning only one row.
}
\examples{
# Return one row
tail1(mtcars)
# Return multiple rows
tail1(x = mtcars, n = 3)

}
\keyword{tail}
