% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_polygons.R
\name{thin_polygons}
\alias{thin_polygons}
\title{Thinning (simplification)}
\usage{
thin_polygons(sf_obj, tolerance)
}
\arguments{
\item{sf_obj}{an objet of class "sf".}

\item{tolerance}{the tolerance value in the metric of the input object (cf.
function `thinnedSpatialPoly`).}
}
\description{
The function performs \code{\link[maptools]{thinnedSpatialPoly}} on a
\code{sf} object of polygons.
}
\examples{
vn <- gadm("vietnam", "sf", 0)
vn2 <- thin_polygons(vn, 0.1)

sp::plot(vn)
sp::plot(vn2)
}
