% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_centroid.R
\name{weighted_centroid}
\alias{weighted_centroid}
\title{Calculate the weighted centroid}
\usage{
weighted_centroid(spdt)
}
\arguments{
\item{spdt}{Spatial Points data frame with weighted information in the
slot \code{data}}
}
\value{
coordinates x,y of a point
}
\description{
Calculate the coordinate of the weighted centroid from a Spatial Points Data
Frame object
}
\examples{
library(sf)

# SpatialPoints
stations <- as(imhen::stations, "Spatial")
stations@data <- stations@data[, "latitude", drop = FALSE]
weighted_centroid(stations)

}
