% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_plot.R
\name{x}
\alias{x}
\title{compat.plot}
\usage{
x
}
\arguments{
\item{x}{data the test is run on.}

\item{y}{data the test is run on.}

\item{FUN}{test function to use.}

\item{test.args}{list of further arguments to be passed on to the test function}

\item{...}{further arguments to be passed on to}

\item{conf.levels}{confidence levels to highlight}

\item{ylim}{range of the vertical axis entered as a vector consisting of lower and upper bound}

\item{xlab}{horizontal axis title}

\item{ylab}{vertical axis title}

\item{precision}{step-width and distance from 1 for compatibility intervals to plot}

\item{digits.x}{significant digits for x-axis labels}

\item{digits.y}{significant digits for y-axis labels}

\item{file.name}{fully qualified file name for graph output file. If not specified, no image file is produced.}
}
\description{
Function which returns a compatibility plot for hypothesis testing in ggplot2 format.
}
\examples{
Instead of writing, for example:
A[A$a==B$a & A$b==B$b & A$c==B$c & A$d==B$d,]
one can write
A[SPSubset(A,B,c("a","b","c","d")),]
}
\keyword{compatibility}
\keyword{ggplot2}
\keyword{hypothesis}
\keyword{plot,}
\keyword{testing,}
