% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_from_raster.R
\name{add_from_raster}
\alias{add_from_raster}
\title{Adds values of a RasterLayer object to a SpatialPointsDataFrame object}
\usage{
add_from_raster(sptsdf, rstr, varname = "new_data")
}
\arguments{
\item{sptsdf}{A \code{SpatialPointsDataFrame} object.}

\item{rstr}{A \code{RasterLayer} object.}

\item{varname}{The name of the new variable. Defaults to "new_data".}
}
\value{
A \code{SpatialPointsDataFrame} object.
}
\description{
Based on the \code{\link[raster]{extract}} function.
}
\examples{
library(sp)

# download vietnam country administrative map in the internal library and in
# the working direction
vn <- sptools::gadm("vietnam", "sp", 0, intlib = TRUE, save = TRUE)

# With a SpatialPointsDataFrame
value <- spsample(vn, 100, "random")
value <- SpatialPointsDataFrame(value, data.frame(variable = 1:100))
value <- add_from_raster(value, srtmVN::getsrtm(), "elevation")

head(value@data)

# With a SpatialPoints
value <- spsample(vn, 100, "random")
value <- SpatialPoints(value, proj4string = vn@proj4string, bbox = vn@bbox)
value <- add_from_raster(value, srtmVN::getsrtm(), "elevation")

head(value@data)
}
