% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pts_by_poly.R
\name{apply_pts_by_poly}
\alias{apply_pts_by_poly}
\title{Apply functions over points by polygons}
\usage{
apply_pts_by_poly(points, polygons, var, f, ...)
}
\arguments{
\item{points}{a `SpatialPointsDataFrame` object.}

\item{polygons}{a `SpatialPolygonsDataFrame` object.}

\item{var}{the name of the variable of `points` on which to apply the
function `f`.}

\item{f}{the function to be applied.}

\item{...}{arguments to passed in `sptools::points_in_polygon`}
}
\description{
This function is a wrapper aroud the `sptools::points_in_polygon` and applies
a function `f` to the variable `var` of the `spatialPointsDataFrame` `points`
object, by the polygons of `spatialPolygonsDataFrame` `polygons` object.
}
\examples{
library(sf)

# SpatialPolygonsDataFrame
vn <- sf::as_Spatial(gadmVN::gadm(level = "province"))
# SpatialPoints
stations <- as(imhen::stations, "Spatial")
names(stations)

# to calculate the mean elevation of the station for each polygon
apply_pts_by_poly(stations, vn, "elevation", mean)
}
\author{
Marc Choisy
}
