% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_list.R
\name{as_list}
\alias{as_list}
\title{Convert a Spatial polygon into a list of Spatial polygons}
\usage{
as_list(plgn)
}
\arguments{
\item{plgn}{a SpatialPolygon*}
}
\description{
Convert a Spatial polygon into a list of Spatial polygons
}
\examples{
library(sf)
library(dplyr)
# download vietnam admin1 administrative map in the internal library and in
# the working direction
vn <- sptools::gadm("vietnam", "sf", 1, intlib = TRUE, save = TRUE)

provinces <- vn \%>\%
  transmute(province = VARNAME_1) \%>\%
  as("Spatial")
class(provinces)
list_provinces <- as_list(provinces)
class(list_provinces)
table(sapply(list_provinces, class))

}
