% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spToolsVectors.R
\name{ci.mean2.vec}
\alias{ci.mean2.vec}
\title{Wrapper for ci.mean2 with vector input}
\usage{
ci.mean2.vec(alpha, m, sd, n)
}
\arguments{
\item{alpha}{Alpha level for 1-alpha confidence}

\item{m}{Numeric vector of length 2: means for groups 1 and 2}

\item{sd}{Numeric vector of length 2: standard deviations for groups 1 and 2}

\item{n}{Numeric vector of length 2: sample sizes for groups 1 and 2}
}
\value{
A 2-row matrix identical to the output of ci.mean2
}
\description{
A wrapper function for ci.mean2 that accepts vectors for means, standard deviations, and sample sizes.
}
\examples{
ci.mean2.vec(.05, c(15.4, 10.3), c(2.67, 2.15), c(30, 20))

}
