% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spToolsReformat.R
\name{ci.tukey.reformat}
\alias{ci.tukey.reformat}
\title{Clean and Reformat Tukey Confidence Interval Output}
\usage{
ci.tukey.reformat(ci_out)
}
\arguments{
\item{ci_out}{A matrix or data frame produced by \code{ci.tukey}, where the first two columns represent pairwise comparisons.}
}
\value{
A cleaned matrix without the first two \code{pair} columns. The rows will be named according to the pairs compared,
using the format \code{"1 v 2"}, \code{"1 v 3"}, etc.
}
\description{
This function takes the output from the Tukey confidence interval function (\code{ci.tukey})
and reformats it by removing the \code{pair} columns and assigning more descriptive row names using the pair indices.
Row names will be formatted as \code{"1 v 2"}, \code{"1 v 3"}, etc.
}
\examples{
out <- ci.tukey(alpha = 0.05, m = c(5, 6, 7), sd = 2, n = 10)
ci.tukey.reformat(out)

}