% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_on_poly.R
\name{crop_on_poly}
\alias{crop_on_poly}
\title{Crops a raster file according to a polygon}
\usage{
crop_on_poly(plgn, rstr)
}
\arguments{
\item{plgn}{spatial polygon sp}

\item{rstr}{raster}
}
\description{
Crops a raster file according to a polygon
}
\examples{
# library(sf)
# library(sptools)
# library(dplyr)
# provinces <- provinces <- "vietnam" \%>\%
#   sptools::gadm("sf", 1) \%>\%
#   transmute(province = VARNAME_1) \%>\%
#   as("Spatial")
# ppp2010 <- worldpopVN::getpop(2010)
# hanoi <- crop_on_poly(subset(provinces, province == "Ha Noi"), ppp2010)

}
