% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_on_poly.R
\name{crop_on_poly}
\alias{crop_on_poly}
\title{Crops a raster file according to a polygon}
\usage{
crop_on_poly(rstr, plgn)
}
\arguments{
\item{rstr}{raster}

\item{plgn}{spatial polygon sp}
}
\description{
Crops a raster file according to a polygon
}
\examples{

library(sf)
library(dplyr)
# download vietnam admin1 administrative map in the internal library and in
# the working direction
vn <- sptools::gadm("vietnam", "sf", 1, intlib = TRUE, save = TRUE)

provinces <- vn \%>\%
  transmute(province = VARNAME_1) \%>\%
  as("Spatial")
srtm <- srtmVN::getsrtm()
hanoi <- crop_on_poly(srtm, subset(provinces, province == "Ha Noi"))

}
