% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{crs}
\alias{crs}
\title{CRS of a Spatial* Object}
\usage{
crs(spobject)
}
\arguments{
\item{spobject}{an object of class Spatial}
}
\value{
A CRS object as defined in the \code{sp} package.
}
\description{
Retrieve the CRS of a Spatial* Object
}
\examples{
vn_prov <- gadmVN::gadm()
crs(vn_prov)
vn_prov2 <- sp::spTransform(vn_prov, projVN)
crs(vn_prov2)
vn_prov3 <- sp::spTransform(vn_prov2, crs(vn_prov))
crs(vn_prov)

}
