% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadm.R
\name{gadm}
\alias{gadm}
\title{Get GADM data}
\usage{
gadm(country, format, level, intlib = TRUE, save = FALSE,
  path = NULL)
}
\arguments{
\item{country}{character string, name of the country to download}

\item{format}{character string, format to downlaod, either sp or sf}

\item{level}{integer between 0 and 3 or 4, level of the administrative
borders, 0 being country borders.}

\item{intlib}{boolean, specifies whether the downloaded file should be saved
in the library of packages. If \code{NULL}, it will be asked interactively.
By default \code{TRUE}, see `Details` for more information.}

\item{save}{boolean, specifies whether the downloaded file should be saved
in a specific path or not. If \code{NULL}, it will be asked interactively.
By default \code{FALSE}, see `Details` for more information.}

\item{path}{character string, path to save the downloaded file. If
\code{NULL}, the file will be saved in the working directory. By default
\code{NULL}, see `Details` for more information.}
}
\description{
Download sp or sf file format from GADM in the internal library of the
package and/or a specified path and/or a temporary path.
}
\details{
If the \code{save} and/or \code{intlib} arguments are \code{NULL}, it will
be asked interactively to the user to decide where to store the file. The
function will return a message if the file is already present in the internal
library or in the specify path. \cr\cr
The table will show you the interaction between these two arguments
to download sp or sf file format from GADM:
}
\examples{

# to download Cambodia country administrative boundaries
kh <- gadm("Cambodia", "sf", 0)

#' # to download Vietnam only in the working directory
vn <- gadm("Vietnam", "sf", 0, intlib = FALSE, save = TRUE)

}
\author{
Marc Choisy, Lucie Contamin
}
