% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadm.R
\name{gadm}
\alias{gadm}
\title{Get GADM data}
\usage{
gadm(country, format, level)
}
\arguments{
\item{country}{character string, name of the country to download}

\item{format}{character string, format to downlaod, either sp or sf}

\item{level}{integer between 0 and 3 or 4, level of the administrative borders,
0 being country borders.}
}
\description{
Download sp or sf file format from GADM
}
\author{
Marc Choisy
}
