% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spTools-package.R
\docType{package}
\name{spTools}
\alias{spTools}
\alias{spTools-package}
\title{spTools: Helper Functions for the statpsych Package}
\description{
A companion package that streamlines data preparation and result formatting for use with the statpsych package.
}
\details{
spTools is a small collection of helper functions that enhance the usability and interpretability
of statistical results from the statpsych package by Douglas Bonett.
Where statpsych provides high-accuracy confidence interval methods for a wide variety of statistics,
spTools handles the pre-processing of data into the required formats and post-processing
the results into more accessible and readable forms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cwendorf/spTools}
  \item Report bugs at \url{https://github.com/cwendorf/spTools/issues}
}

}
\author{
\strong{Maintainer}: Craig Wendorf
\email{cwendorf@uwsp.edu}
}
