% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spTools-package.R
\docType{package}
\name{spTools}
\alias{spTools}
\alias{spTools-package}
\title{spTools: Additional Functions for the statpsych Package}
\description{
spTools is a small collection of functions that add to the array of tools in Doug Bonetts statpsych package.
Some functions provide calculations not present in statpsych, whereas other functions offer vectorized versions of statpsych functions.
Other functions attempt to further standardize statpsych output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cwendorf/spTools}
  \item Report bugs at \url{https://github.com/cwendorf/spTools/issues}
}

}
\author{
\strong{Maintainer}: Craig Wendorf
\email{cwendorf@uwsp.edu}
}
